ULong -> UInt32
git-svn-id: https://svn.code.sf.net/p/gexperts/code@4049 03fc42d2-69ba-4c40-8a67-5bfb646c2f2c

+++ trunk/Source/UsesExpert/GX_UnitExportList.pas
@@ -126,7 +126,7 @@ class function TUnitIdentifier.FromLine(_s: PAnsiChar): TUnitIdentifier;
 var
   Parts: array of PAnsiChar;
   Idx: Integer;
-  l: ulong;
+  l: UInt32;
   LLineNo: Integer;
   p: PAnsiChar;
   Start: PAnsiChar;
